rename equilibrate.py->equilibration.py
